fix(frontier): use pointer-width GMP mp_limb_t (Windows LLP64)
The Windows CI lane failed to compile frontier_register_value_conversions_
owned.zig: "@ptrCast increases pointer alignment ('[*c]c_ulong' has
alignment '4', '[*]u64' has alignment '8')" at the raw GMP limb reads.

Root cause: the owners defined the GMP limb type as `c_ulong`. On Win64
(LLP64) `unsigned long` is 32-bit, but GMP there uses 64-bit limbs
(mp_limb_t = unsigned long long), so `_mp_d` was typed [*c]u32 (align 4)
and could not @ptrCast to [*]u64 (align 8). It happened to compile on
Linux x64 (c_ulong = 64-bit) and ARM32 firmware (c_ulong = 32-bit, limbs
32-bit), so local builds were green.

The limb width equals the pointer width on every target z47 builds
(64-bit on Linux and Win64, 32-bit on ARM32), so define mp_limb_t = usize
in the three owners that model mpz_struct (registerValueConversions,
integers, dateTime). This matches GMP's actual limb size on all three and
keeps the is32 (= pointer-width) branch selection consistent.

No-op on Linux/ARM32 (usize is identical to the previous c_ulong there).
Verified: x86_64-windows-gnu cross-compile of the frontier object now
succeeds; zig build sim; zig build test (9530 + 2 host); dmcp dmcp_pkgs_all
(ARM32 is32 path, all packages fit); boundary guard.

+++ zig_src/frontier/frontier_date_time_owned.zig
@@ -50,8 +50,9 @@ const realContext_t = extern struct {
 const calcRegister_t = i16;
 const angularMode_t = c_int;
 
-// GMP mpz_struct. mp_limb_t is c_ulong (32-bit on arm32 firmware, 64-bit on host).
-const mp_limb_t = c_ulong;
+// GMP mpz_struct. The limb width == pointer width on every z47 target (64-bit on
+// Linux and Win64, 32-bit on ARM32 firmware) — NOT c_ulong, which is 32-bit on Win64.
+const mp_limb_t = usize;
 const mpz_struct = extern struct {
     _mp_alloc: c_int,
     _mp_size: c_int,
