+++ zig_src/frontier/frontier_integers_owned.zig
@@ -18,9 +18,10 @@ const extra_info: bool = frontier_build_options.extra_info_on_calc_error;
 
 // ---------------------------------------------------------------------------
 // GMP long integer (mpz). The mpz_* names are header macros; the real symbols are
-// __gmpz_*. mp_limb_t is c_ulong (32-bit on arm32 firmware, 64-bit on host).
+// __gmpz_*. The limb width == pointer width on every z47 target (64-bit on Linux
+// and Win64, 32-bit on ARM32 firmware) — NOT c_ulong, which is 32-bit on Win64.
 // ---------------------------------------------------------------------------
-const mp_limb_t = c_ulong;
+const mp_limb_t = usize;
 const mpz_struct = extern struct {
     _mp_alloc: c_int,
     _mp_size: c_int,
